+++ src/PolygonClipper/Box2.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Six Labors.
 // Licensed under the Six Labors Split License.
 
+using System;
 using System.Runtime.CompilerServices;
 
 namespace SixLabors.PolygonClipper;
@@ -32,6 +33,13 @@ public readonly struct Box2 : IEquatable<Box2>
     }
 
     /// <summary>
+    /// Gets an invalid bounds instance.
+    /// </summary>
+    public static Box2 Invalid { get; } = new(
+        new Vertex(double.MaxValue, double.MaxValue),
+        new Vertex(-double.MaxValue, -double.MaxValue));
+
+    /// <summary>
     /// Gets the minimum xy-coordinate.
     /// </summary>
     public Vertex Min { get; }
@@ -60,6 +68,44 @@ public readonly struct Box2 : IEquatable<Box2>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static bool operator !=(in Box2 left, in Box2 right)
         => !(left == right);
+
+    /// <summary>
+    /// Returns true if the box is empty.
+    /// </summary>
+    /// <returns><see langword="true"/> if the box is empty; otherwise, <see langword="false"/>.</returns>
+    public bool IsEmpty() => this.Max.X <= this.Min.X || this.Max.Y <= this.Min.Y;
+
+    /// <summary>
+    /// Returns true if the point lies within the box.
+    /// </summary>
+    /// <param name="point">The point to test.</param>
+    /// <returns><see langword="true"/> if the point lies within the box; otherwise, <see langword="false"/>.</returns>
+    public bool Contains(in Vertex point)
+        => point.X > this.Min.X && point.X < this.Max.X && point.Y > this.Min.Y && point.Y < this.Max.Y;
+
+    /// <summary>
+    /// Returns true if the box contains another box.
+    /// </summary>
+    /// <param name="bounds">The other box.</param>
+    /// <returns><see langword="true"/> if the box contains the other box; otherwise, <see langword="false"/>.</returns>
+    public bool Contains(in Box2 bounds)
+        => bounds.Min.X >= this.Min.X && bounds.Max.X <= this.Max.X &&
+           bounds.Min.Y >= this.Min.Y && bounds.Max.Y <= this.Max.Y;
+
+    /// <summary>
+    /// Returns true if the boxes intersect.
+    /// </summary>
+    /// <param name="bounds">The other box.</param>
+    /// <returns><see langword="true"/> if the boxes intersect; otherwise, <see langword="false"/>.</returns>
+    public bool Intersects(in Box2 bounds)
+        => Math.Max(this.Min.X, bounds.Min.X) <= Math.Min(this.Max.X, bounds.Max.X) &&
+           Math.Max(this.Min.Y, bounds.Min.Y) <= Math.Min(this.Max.Y, bounds.Max.Y);
+
+    /// <summary>
+    /// Returns the midpoint of the box.
+    /// </summary>
+    /// <returns>The midpoint.</returns>
+    public Vertex MidPoint() => new((this.Min.X + this.Max.X) / 2D, (this.Min.Y + this.Max.Y) / 2D);
 
     /// <summary>
     /// Adds another bounding box to this instance.
